+++ db/migrate/20141227120907_add_export_to_user.rb
@@ -1,0 +1,7 @@
+class AddExportToUser < ActiveRecord::Migration
+  def change
+    add_column :users, :export, :string
+    add_column :users, :exported_at, :datetime
+    add_column :users, :exporting, :boolean, default: false
+  end
+end